+++ example/lib/calendar_header_title.dart
@@ -3,10 +3,7 @@ import 'package:flutter/material.dart';
 class CalendarHeaderTitle extends StatelessWidget {
   final String displayMonth;
 
-  const CalendarHeaderTitle({
-    required this.displayMonth,
-    super.key,
-  });
+  const CalendarHeaderTitle({required this.displayMonth, super.key});
 
   @override
   Widget build(BuildContext context) {